# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gerenciamento de fazendas leiteiras\nSistema completo para gestão de animais, coletas de leite, reproduções, vendas e dívidas",
        "title": "FazendaPro API",
        "contact": {
            "name": "Suporte FazendaPro",
            "email": "suporte@fazendapro.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/milk-collections": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra uma nova coleta de leite de um animal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "milk-collections"
                ],
                "summary": "Criar coleta de leite",
                "parameters": [
                    {
                        "description": "Dados da coleta",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateMilkCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.MilkCollectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/milk-collections/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de uma coleta de leite existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "milk-collections"
                ],
                "summary": "Atualizar coleta de leite",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da coleta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateMilkCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MilkCollectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AnimalData": {
            "description": "Dados principais de um animal na fazenda",
            "type": "object",
            "properties": {
                "animal_name": {
                    "description": "Nome do animal",
                    "type": "string",
                    "example": "Branquinha"
                },
                "animal_type": {
                    "description": "Tipo de animal (enum)",
                    "type": "integer",
                    "example": 1
                },
                "birth_date": {
                    "description": "Data de nascimento (ISO format)",
                    "type": "string",
                    "example": "2020-01-15"
                },
                "breed": {
                    "description": "Raça",
                    "type": "string",
                    "example": "Holandesa"
                },
                "castrated": {
                    "description": "Se está castrado",
                    "type": "boolean",
                    "example": false
                },
                "confinement": {
                    "description": "Se está em confinamento",
                    "type": "boolean",
                    "example": false
                },
                "current_batch": {
                    "description": "Lote atual",
                    "type": "integer",
                    "example": 1
                },
                "ear_tag_number_local": {
                    "description": "Número da brinco local",
                    "type": "integer",
                    "example": 123
                },
                "ear_tag_number_register": {
                    "description": "Número de registro",
                    "type": "integer",
                    "example": 456
                },
                "farm_id": {
                    "description": "ID da fazenda",
                    "type": "integer",
                    "example": 1
                },
                "father_id": {
                    "description": "ID do pai (opcional)",
                    "type": "integer",
                    "example": 10
                },
                "fertilization": {
                    "description": "Se está fertilizada",
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "description": "ID único do animal",
                    "type": "integer",
                    "example": 1
                },
                "mother_id": {
                    "description": "ID da mãe (opcional)",
                    "type": "integer",
                    "example": 20
                },
                "photo": {
                    "description": "Foto em base64",
                    "type": "string",
                    "example": "data:image/jpeg;base64,..."
                },
                "purpose": {
                    "description": "Propósito do animal (enum)",
                    "type": "integer",
                    "example": 1
                },
                "sex": {
                    "description": "Sexo (1=Fêmea, 2=Macho)",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "description": "Status do animal (enum)",
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "description": "Tipo de animal",
                    "type": "string",
                    "example": "Bovino"
                }
            }
        },
...
        "handlers.CreateMilkCollectionRequest": {
            "description": "Dados necessários para criar uma nova coleta de leite",
            "type": "object",
            "required": [
                "animal_id",
                "date",
                "liters"
            ],
            "properties": {
                "animal_id": {
                    "description": "ID do animal",
                    "type": "integer",
                    "example": 1
                },
                "date": {
                    "description": "Data da coleta (YYYY-MM-DD)",
                    "type": "string",
                    "example": "2024-01-15"
                },
                "liters": {
                    "description": "Quantidade de litros",
                    "type": "number",
                    "minimum": 0,
                    "example": 25.5
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.MilkCollectionData": {
            "type": "object",
            "properties": {
                "animal": {
                    "$ref": "#/definitions/handlers.AnimalData"
                },
                "animal_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "liters": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "handlers.MilkCollectionResponse": {
            "description": "Resposta com dados de uma coleta de leite",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Dados da coleta",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handlers.MilkCollectionData"
                        }
                    ]
                },
                "message": {
                    "description": "Mensagem de resposta",
                    "type": "string"
                },
                "success": {
                    "description": "Indica sucesso",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e depois o token JWT. Exemplo: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}